* commit '64208a43420173c31fd3253483bfc51eb4dab8bc':
  SM-415 ~ Fixed allow actions for person
  SM-415 + Added executing poll
  SM-415 + Added editing poll

+++ src/app/module/person/person-head/person-head/person-head.component.ts
@@ -1,26 +1,33 @@
-import {Component, EventEmitter, Input, OnInit, Output} from '@angular/core';
+import {Component, EventEmitter, Input, OnDestroy, OnInit, Output} from '@angular/core';
 import {Person} from '../../../../data/remote/model/person';
 import {MenuItem} from '../../../common/item-line/model/menu-item';
 import {ImageType} from '../../../../data/remote/model/file/image/image-type';
 import {FileClass} from '../../../../data/remote/model/file/base/file-class';
 import {FileApiService} from '../../../../data/remote/rest-api/api/file/file-api.service';
-import {map} from 'rxjs/operators';
+import {map, takeWhile} from 'rxjs/operators';
 import {PersonApiService} from '../../../../data/remote/rest-api/api/person/person-api.service';
 import {Router} from '@angular/router';
 import {ParticipantRestApiService} from '../../../../data/remote/rest-api/participant-rest-api.service';
+import {MatIconRegistry} from '@angular/material';
+import {DomSanitizer} from '@angular/platform-browser';
+import {AuthorizationService} from '../../../../shared/authorization.service';
+import {PermissionService} from '../../../../shared/permission.service';
 
 @Component({
   selector: 'app-person-head',
   templateUrl: './person-head.component.html',
   styleUrls: ['./person-head.component.scss']
 })
-export class PersonHeadComponent implements OnInit {
+export class PersonHeadComponent implements OnInit, OnDestroy {
 
   @Input()
   public person: Person;
 
   @Input()
   public canEdit: boolean;
+
+  @Input()
+  public canEditPerson = true;
 
   @Output()
   public readonly navigate = new EventEmitter();
@@ -32,32 +39,60 @@ export class PersonHeadComponent implements OnInit {
   public readonly fileClassClass = FileClass;
   public actions: MenuItem[] = [];
   public hasConnection: boolean;
+  private _notDestroyed = true;
 
   constructor(private _fileApiService: FileApiService,
               private _router: Router,
               private _participantRestApiService: ParticipantRestApiService,
+              private _matIconRegistry: MatIconRegistry,
+              private _domSanitizer: DomSanitizer,
+              private _authorizationService: AuthorizationService,
+              private _permissionService: PermissionService,
               private _personApiService: PersonApiService) {
+    this._matIconRegistry.addSvgIcon('personDelete', this._domSanitizer.bypassSecurityTrustResourceUrl('assets/img/person_delete.svg'));
   }
 
   public async ngOnInit(): Promise<void> {
     this.hasConnection = (await this._participantRestApiService.hasConnection({id: this.person.id})).value;
 
-    this.actions = [{
-      translationLabel: 'sendMessage', action: async item => {
-        await this.onSendMessage();
-      }
-    }, {
-      translationLabel: this.hasConnection ? 'removeContact' : 'addContact',
-      action: async item => {
-        this.hasConnection = (await this._participantRestApiService.hasConnection({id: this.person.id})).value;
-        if (this.hasConnection) {
-          await this._participantRestApiService.removeConnection({id: this.person.id});
-        } else {
-          await this._participantRestApiService.createConnection({id: this.person.id});
+    this._authorizationService.person$
+      .pipe(takeWhile(() => this._notDestroyed))
+      .subscribe(person => {
+        const canEditPerson = this._permissionService.canEditPerson(this.person).toPromise();
+        this.actions = [];
+        if (this.canEditPerson && canEditPerson) {
+          this.actions.push({
+            iconName: 'edit', action: async item => {
+              this.editPerson.emit();
+            }
+          });
         }
-        item.translationLabel = this.hasConnection ? 'removeContact' : 'addContact';
-      }
-    }];
+
+        if (!this._permissionService.areYouCreator(this.person, person)) {
+          this.actions.push(...[
+            {
+              iconName: 'message', action: async item => {
+                await this.onSendMessage();
+              }
+            }, {
+              iconName: this.hasConnection ? 'personDelete' : 'person_add',
+              action: async item => {
+                this.hasConnection = (await this._participantRestApiService.hasConnection({id: this.person.id})).value;
+                if (this.hasConnection) {
+                  await this._participantRestApiService.removeConnection({id: this.person.id});
+                } else {
+                  await this._participantRestApiService.createConnection({id: this.person.id});
+                }
+                item.iconName = this.hasConnection ? 'personDelete' : 'person_add';
+              }
+            }
+          ]);
+        }
+      });
+  }
+
+  public ngOnDestroy(): void {
+    delete this._notDestroyed;
   }
 
   public async onSendMessage() {
